Refactor: reusable VendorDependencies object
This pulls the vendoring logic out of the bundler updater, so we can
reuse it for other ecosystems as well.
