+++ bundler/lib/dependabot/bundler/file_updater.rb
@@ -2,6 +2,7 @@
 
 require "dependabot/file_updaters"
 require "dependabot/file_updaters/base"
+require "dependabot/file_updaters/vendor_updater"
 
 module Dependabot
   module Bundler
@@ -21,6 +22,7 @@ module Dependabot
       end
 
       # rubocop:disable Metrics/PerceivedComplexity
+      # rubocop:disable Metrics/AbcSize
       def updated_dependency_files
         updated_files = []
 
@@ -54,13 +56,16 @@ module Dependabot
         check_updated_files(updated_files)
 
         base_dir = updated_files.first.directory
-        updated_vendor_cache_files(base_directory: base_dir).each do |file|
+        vendor_updater.
+          updated_vendor_cache_files(base_directory: base_dir).
+          each do |file|
           updated_files << file
         end
 
         updated_files
       end
       # rubocop:enable Metrics/PerceivedComplexity
+      # rubocop:enable Metrics/AbcSize
 
       private
 
@@ -76,69 +81,11 @@ module Dependabot
           end
       end
 
-      # Returns changed files in the vendor/cache folder
-      #
-      # @param base_directory [String] Update config base directory
-      # @return [Array<Dependabot::DependencyFile>]
-      def updated_vendor_cache_files(base_directory:)
-        return [] unless repo_contents_path && vendor_cache_dir
-
-        Dir.chdir(repo_contents_path) do
-          relative_dir = vendor_cache_dir.sub("#{repo_contents_path}/", "")
-          status = SharedHelpers.run_shell_command(
-            "git status --untracked-files=all --porcelain=v1 #{relative_dir}"
-          )
-          changed_paths = status.split("\n").map { |l| l.split(" ") }
-          changed_paths.map do |type, path|
-            deleted = type == "D"
-            encoding = ""
-            encoded_content = File.read(path) unless deleted
-            if binary_file?(path)
-              encoding = Dependabot::DependencyFile::ContentEncoding::BASE64
-              encoded_content = Base64.encode64(encoded_content) unless deleted
-            end
-            Dependabot::DependencyFile.new(
-              name: path,
-              content: encoded_content,
-              directory: base_directory,
-              deleted: deleted,
-              content_encoding: encoding
-            )
-          end
-        end
-      end
-
-      # notable filenames without a reliable extension:
-      TEXT_FILE_NAMES = [
-        "Gemfile",
-        "Gemfile.lock",
-        ".bundlecache",
-        ".gitignore"
-      ].freeze
-
-      TEXT_FILE_EXTS = [
-        # code
-        ".rb",
-        ".erb",
-        ".gemspec",
-        ".js",
-        ".html",
-        # config
-        ".json",
-        ".xml",
-        ".toml",
-        ".yaml",
-        ".yml",
-        # docs
-        ".md",
-        ".txt"
-      ].freeze
-
-      def binary_file?(path)
-        return false if TEXT_FILE_NAMES.include?(File.basename(path))
-        return false if TEXT_FILE_EXTS.include?(File.extname(path))
-
-        true
+      def vendor_updater
+        Dependabot::FileUpdaters::VendorUpdater.new(
+          repo_contents_path: repo_contents_path,
+          vendor_dir: vendor_cache_dir
+        )
       end
 
       def check_required_files
